현우

+++ TEAM3FINAL/HHW/FrmBORManage.cs
@@ -159,7 +159,7 @@ namespace TEAM3FINAL
                     frm.BOR_PROCS_TIME = Convert.ToInt32(dgvBORList.CurrentRow.Cells[8].Value);
                     frm.BOR_PRIORT = Convert.ToInt32(dgvBORList.CurrentRow.Cells[9].Value);
                     frm.BOR_PROCS_LEADTIME = Convert.ToInt32(dgvBORList.CurrentRow.Cells[10].Value);
-                    frm.BOR_YIELD = Convert.ToInt32(dgvBORList.CurrentRow.Cells[11].Value);
+                    frm.BOR_YIELD = Convert.ToDecimal(dgvBORList.CurrentRow.Cells[11].Value);
                     frm.BOR_USE_YN = dgvBORList.CurrentRow.Cells[12].Value.ToString();
                     frm.BOR_REMARK = dgvBORList.CurrentRow.Cells[13].Value.ToString();
                     frm.ShowDialog();
